refactor: align with PR #420 pattern — shared insights module
Move types to src/models/maestro/insights.types.ts, extract shared
fetchInstanceStatusByDate() to src/services/maestro/insights.ts,
use barrel exports, and update all imports to match PR #420 structure.
