+++ src/models/common/index.ts
@@ -1,3 +1,2 @@
 export * from './request-spec';
 export * from './types';
-export * from './insights.types';
