+++ src/services/maestro/processes/index.ts
@@ -39,5 +39,3 @@ export * from '../../../models/maestro/process-instances.models';
 export * from '../../../models/maestro/process-incidents.types';
 export * from '../../../models/maestro/process-incidents.models';
 export * from '../../../models/maestro/insights.types';
-export { TimeSliceUnit } from '../../../models/common';
-export type { MaestroInsightsOptions, InstanceStatusByDateResponse } from '../../../models/common';
